Add files via upload
This resource for ESX adds possibility for different jobs to send bills to players, for example making police units able to give people fines. It comes with a menu for paying bills, to open the menu the default keybind is F7.

There is a developer server event available in order to register bills in the database, see default resources for examples.